+++ src/components/Workspace/Passwords/FoldersTree/FoldersTree.js
@@ -108,6 +108,7 @@ class FoldersTree extends React.Component {
     const openFolders = this.state.openFolders;
     openFolders.push(folder);
     this.setState({openFolders});
+    Plugin.request('passbolt.plugin.folders.update-local-storage');
   }
 
   /**
@@ -115,6 +116,7 @@ class FoldersTree extends React.Component {
    */
   handleClickOnTitle() {
     this.props.onSelectRoot();
+    Plugin.request('passbolt.plugin.folders.update-local-storage');
   }
 
   /**
@@ -147,6 +149,9 @@ class FoldersTree extends React.Component {
    */
   handleSectionTitleClickCaretEvent() {
     const open = !this.state.open;
+    if (open) {
+      Plugin.request('passbolt.plugin.folders.update-local-storage');
+    }
     this.setState({open});
   }
 
@@ -235,6 +240,7 @@ class FoldersTree extends React.Component {
    * @param {Object} folder The folder
    */
   handleFolderSelectEvent(event, folder) {
+    Plugin.request('passbolt.plugin.folders.update-local-storage');
     this.props.onSelect(folder);
   }
 
@@ -434,7 +440,7 @@ class FoldersTree extends React.Component {
     }
 
     return (
-      <div className="folders navigation first accordion">
+      <div className="folders navigation accordion">
         {this.renderDragFeedback()}
         <div className="accordion-header1">
           <div className={`${isOpen ? "open" : "close"} node root`}>